+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## New in git-machete 3.14.0
 
 - added: `push=no` and `rebase=no` branch qualifiers that control push and rebase behaviour of `git machete traverse`
+- added: `machete.github.domain` config key to support GitHub Enterprise domains
 
 ## New in git-machete 3.13.2